Proyecto_IA_Retro/requirements.txt
fastapi
uvicorn
requests
faiss-cpu
sentence-transformers
watchdog
ollama
